// src/components/FixedGrid/types.ts
/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
export enum GutterSize {
  /**
   * 24px
   */
  XXLARGE = 'xxLarge',

  /**
   * 20px
   */
  XLARGE = 'xLarge',
  /**
   * 16px
   */
  LARGE = 'large',
  /**
   * 12px
   */
  MEDIUM = 'medium',
  /**
   * 8px
   */
  SMALL = 'small',
  /**
   * 4px
   */
  XSMALL = 'xSmall',
  /**
   * 0px
   */
  NONE = 'none',
}